WINDUP-2627 - improvements for detecting user-specified RHAMT cli location preference
